fix: Necromancer patrol issue
Fixed so necromancer receives correct datatype on instantiation which
resolves the issue where necromancer does not patrol.

+++ core/src/main/java/com/mygdx/platformer/characters/enemies/Necromancer.java
@@ -35,7 +35,7 @@ public class Necromancer extends BaseEnemy {
             AppConfig.NECROMANCER_HEIGHT);
     }
 
-    public Necromancer(World world, Vector2 position, int health, int speed) {
+    public Necromancer(World world, Vector2 position, int health, float speed) {
         super(world, position, health, speed, AppConfig.NECROMANCER_WIDTH,
             AppConfig.NECROMANCER_HEIGHT);
     }
